+++ features/step_definitions/factory_steps.rb
@@ -48,6 +48,12 @@ Given /^a page "([^"]+)" with title "(.+)"$/ do |path, title|
   create :page, path: path, title: title, body: "<HTML>BODY</HTML>"
 end
 
+Given /^a "(\w+)" page with title "(.+)" and category "(.+)"$/ do |published, title, category|
+  category_obj = Landable::Category.where('lower(name) = ?', category.downcase).first
+  @page = create :page, title: title, category: category_obj
+  @page.publish! author: create(:author) if published == 'published'
+end
+
 Given 'page "$path" redirects to "$url" with status $code' do |path, url, code|
   page = create :page, :redirect, path: path, redirect_url: url, status_code: code, body: "BODY"
   page.publish! author: create(:author)